[EGD-7217] Separate commonly used headers from application-calendar
Move dateCommon.hpp to module-utils/time
