+++ module-apps/application-calendar/include/application-calendar/data/CalendarData.hpp
@@ -3,10 +3,10 @@
 
 #pragma once
 
-#include "dateCommon.hpp"
 #include <application-calendar/widgets/CalendarStyle.hpp>
 
 #include <module-gui/gui/SwitchData.hpp>
+#include <time/dateCommon.hpp>
 
 enum class EventAction
 {
