+++ module-apps/apps-common/widgets/DateWidget.cpp
@@ -8,7 +8,7 @@
 #include <time/time_conversion.hpp>
 #include <time/time_date_validation.hpp>
 #include <module-gui/gui/input/InputEvent.hpp>
-#include <application-calendar/data/dateCommon.hpp>
+#include <time/dateCommon.hpp>
 
 namespace date_and_time = style::window::date_and_time;
