+++ module-db/Tables/AlarmsTable.hpp
@@ -7,7 +7,7 @@
 #include "Database/Database.hpp"
 #include "Record.hpp"
 #include "Table.hpp"
-#include <application-calendar/data/dateCommon.hpp>
+#include <time/dateCommon.hpp>
 #include <utf8/UTF8.hpp>
 
 struct AlarmsRecord;
