+++ module-services/service-alarm/include/service-alarm/AlarmMessage.hpp
@@ -3,11 +3,12 @@
 
 #pragma once
 
-#include <application-calendar/data/dateCommon.hpp>
 #include <module-db/Interface/AlarmEventRecord.hpp>
 #include <module-db/Interface/EventRecord.hpp>
 #include <MessageType.hpp>
 #include <Service/Message.hpp>
+
+#include <time/dateCommon.hpp>
 
 #include <vector>
