+++ module-utils/ical/ParserICS.hpp
@@ -2,8 +2,11 @@
 // For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
 
 #pragma once
-#include <application-calendar/data/dateCommon.hpp>
+
+#include <time/dateCommon.hpp>
+
 #include <json11.hpp>
+
 #include <memory>
 
 /**
