+++ module-utils/time/CMakeLists.txt
@@ -18,6 +18,8 @@ target_sources(time
         time/TimeRangeParser.cpp
         time/TimeZone.cpp
         time/time_conversion_factory.cpp
+
+        time/dateCommon.hpp
 )
 
 target_include_directories(time PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>)
