+++ module-utils/time/test/unittest_TimeRangeParser.cpp
@@ -3,8 +3,9 @@
 
 #define CATCH_CONFIG_MAIN
 #include <catch2/catch.hpp>
-#include "time/TimeRangeParser.hpp"
-#include <application-calendar/data/dateCommon.hpp>
+
+#include <time/TimeRangeParser.hpp>
+#include <time/dateCommon.hpp>
 
 #include <cstdint>
 #include <cstring>
