+++ module-utils/time/time/FromTillDate.hpp
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include <application-calendar/data/dateCommon.hpp>
+#include "dateCommon.hpp"
 
 namespace utils
 {
